renamed to ThemeToggle.vue
